+++ sqlite_create_tables.sql
@@ -1,7 +1,7 @@
 CREATE TABLE book
 (
 	id		 VARCHAR(32) NOT NULL PRIMARY KEY,
-	myoid		 INTEGER NOT NULL,
+	myoid		 BIGINT NOT NULL,
 	title		 TEXT NOT NULL,
 	edition		 VARCHAR(8) NOT NULL,
 	author		 TEXT NOT NULL,
@@ -30,8 +30,8 @@ CREATE TABLE book
 
 CREATE TABLE book_copy_info
 (
-	item_oid	 INTEGER NOT NULL,
-	myoid		 INTEGER NOT NULL,
+	item_oid	 BIGINT NOT NULL,
+	myoid		 BIGINT NOT NULL,
 	copyid		 VARCHAR(64) NOT NULL,
 	copy_number	 INTEGER NOT NULL DEFAULT 1,
 	PRIMARY KEY(item_oid, copyid)
@@ -46,7 +46,7 @@ END;
 CREATE TABLE cd
 (
 	id		 VARCHAR(32) NOT NULL PRIMARY KEY,
-	myoid		 INTEGER NOT NULL,
+	myoid		 BIGINT NOT NULL,
 	title		 TEXT NOT NULL,
 	artist		 TEXT NOT NULL,
 	recording_label	 TEXT NOT NULL,
@@ -73,7 +73,7 @@ CREATE TABLE cd
 
 CREATE TABLE cd_songs
 (
-	item_oid	 INTEGER NOT NULL,
+	item_oid	 BIGINT NOT NULL,
 	albumnum	 INTEGER NOT NULL DEFAULT 1,
 	songnum		 INTEGER NOT NULL DEFAULT 1,
 	songtitle	 VARCHAR(256) NOT NULL,
@@ -83,8 +83,8 @@ CREATE TABLE cd_songs
 
 CREATE TABLE cd_copy_info
 (
-	item_oid	 INTEGER NOT NULL,
-	myoid		 INTEGER NOT NULL,
+	item_oid	 BIGINT NOT NULL,
+	myoid		 BIGINT NOT NULL,
 	copyid		 VARCHAR(64) NOT NULL,
 	copy_number	 INTEGER NOT NULL DEFAULT 1,
 	PRIMARY KEY(item_oid, copyid)
@@ -100,7 +100,7 @@ END;
 CREATE TABLE dvd
 (
 	id		 VARCHAR(32) NOT NULL PRIMARY KEY,
-	myoid		 INTEGER NOT NULL,
+	myoid		 BIGINT NOT NULL,
 	title		 TEXT NOT NULL,
 	studio		 TEXT NOT NULL,
 	rdate		 VARCHAR(32) NOT NULL,
@@ -129,8 +129,8 @@ CREATE TABLE dvd
 
 CREATE TABLE dvd_copy_info
 (
-	item_oid	 INTEGER NOT NULL,
-	myoid		 INTEGER NOT NULL,
+	item_oid	 BIGINT NOT NULL,
+	myoid		 BIGINT NOT NULL,
 	copyid		 VARCHAR(64) NOT NULL,
 	copy_number	 INTEGER NOT NULL DEFAULT 1,
 	PRIMARY KEY(item_oid, copyid)
@@ -145,7 +145,7 @@ END;
 CREATE TABLE magazine
 (
 	id		 VARCHAR(32) NOT NULL,
-	myoid		 INTEGER NOT NULL,
+	myoid		 BIGINT NOT NULL,
 	title		 TEXT NOT NULL,
 	pdate		 VARCHAR(32) NOT NULL,
 	publisher	 TEXT NOT NULL,
@@ -172,8 +172,8 @@ CREATE TABLE magazine
 
 CREATE TABLE magazine_copy_info
 (
-	item_oid	 INTEGER NOT NULL,
-	myoid		 INTEGER NOT NULL,
+	item_oid	 BIGINT NOT NULL,
+	myoid		 BIGINT NOT NULL,
 	copyid		 VARCHAR(64) NOT NULL,
 	copy_number	 INTEGER NOT NULL DEFAULT 1,
 	PRIMARY KEY(item_oid, copyid)
@@ -188,7 +188,7 @@ END;
 CREATE TABLE videogame
 (
 	id		 VARCHAR(32) NOT NULL PRIMARY KEY,
-	myoid		 INTEGER NOT NULL,
+	myoid		 BIGINT NOT NULL,
 	title		 TEXT NOT NULL,
 	developer	 TEXT NOT NULL,
 	genre		 VARCHAR(64) NOT NULL,
@@ -213,8 +213,8 @@ CREATE TABLE videogame
 
 CREATE TABLE videogame_copy_info
 (
-	item_oid	 INTEGER NOT NULL,
-	myoid		 INTEGER NOT NULL,
+	item_oid	 BIGINT NOT NULL,
+	myoid		 BIGINT NOT NULL,
 	copyid		 VARCHAR(64) NOT NULL,
 	copy_number	 INTEGER NOT NULL DEFAULT 1,
 	PRIMARY KEY(item_oid, copyid)
@@ -228,7 +228,7 @@ END;
 
 CREATE TABLE book_borrower
 (
-	item_oid	 INTEGER NOT NULL,
+	item_oid	 BIGINT NOT NULL,
 	memberid	 VARCHAR(16) NOT NULL,
 	reserved_date	 VARCHAR(32) NOT NULL,
 	duedate		 VARCHAR(32) NOT NULL,
@@ -240,7 +240,7 @@ CREATE TABLE book_borrower
 
 CREATE TABLE cd_borrower
 (
-	item_oid	 INTEGER NOT NULL,
+	item_oid	 BIGINT NOT NULL,
 	memberid	 VARCHAR(16) NOT NULL,
 	reserved_date	 VARCHAR(32) NOT NULL,
 	duedate		 VARCHAR(32) NOT NULL,
@@ -252,7 +252,7 @@ CREATE TABLE cd_borrower
 
 CREATE TABLE dvd_borrower
 (
-	item_oid	 INTEGER NOT NULL,
+	item_oid	 BIGINT NOT NULL,
 	memberid	 VARCHAR(16) NOT NULL,
 	reserved_date	 VARCHAR(32) NOT NULL,
 	duedate		 VARCHAR(32) NOT NULL,
@@ -264,7 +264,7 @@ CREATE TABLE dvd_borrower
 
 CREATE TABLE magazine_borrower
 (
-	item_oid	 INTEGER NOT NULL,
+	item_oid	 BIGINT NOT NULL,
 	memberid	 VARCHAR(16) NOT NULL,
 	reserved_date	 VARCHAR(32) NOT NULL,
 	duedate		 VARCHAR(32) NOT NULL,
@@ -276,7 +276,7 @@ CREATE TABLE magazine_borrower
 
 CREATE TABLE videogame_borrower
 (
-	item_oid	 INTEGER NOT NULL,
+	item_oid	 BIGINT NOT NULL,
 	memberid	 VARCHAR(16) NOT NULL,
 	reserved_date	 VARCHAR(32) NOT NULL,
 	duedate		 VARCHAR(32) NOT NULL,
@@ -305,7 +305,7 @@ CREATE TABLE member
 CREATE TABLE member_history
 (
 	memberid	 VARCHAR(16) NOT NULL,
-	item_oid	 INTEGER NOT NULL,
+	item_oid	 BIGINT NOT NULL,
 	copyid		 VARCHAR(64) NOT NULL,
 	reserved_date	 VARCHAR(32) NOT NULL,
 	duedate		 VARCHAR(32) NOT NULL,
